requirements.txt
pandas==2.1.4
numpy==1.26.2
faker==20.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
pytest-cov==4.1.0
loguru==0.7.2